requirements.txt
# Python packages needed for this project
# Install with: pip install -r requirements.txt

# Data manipulation
pandas>=2.0.0           # For reading CSV and processing data
numpy>=1.24.0           # For numerical operations

# MySQL connection
mysql-connector-python>=8.0.0   # Connect Python to MySQL

# Environment variables (for storing passwords safely)
python-dotenv>=1.0.0    # Read .env file

# Visualization (for Python dashboards)
matplotlib>=3.7.0       # For creating charts
seaborn>=0.12.0         # For beautiful visualizations

# Optional: Uncomment if you want Jupyter notebooks
# jupyter>=1.0.0
# plotly>=5.0.0         # For interactive charts
